CB-1095 Rename rdb to redbeams
