--- rdb/src/main/java/com/sequenceiq/rdb/api/RdbApi.java
@@ -1,10 +1,0 @@
-package com.sequenceiq.rdb.api;
-
-public class RdbApi {
-
-    public static final String API_ROOT_CONTEXT = "/api";
-
-    private RdbApi() {
-    }
-}
-
