--- rdb/src/main/java/com/sequenceiq/rdb/api/model/ExceptionResponse.java
@@ -1,18 +1,0 @@
-package com.sequenceiq.rdb.api.model;
-
-public class ExceptionResponse {
-
-    private String message;
-
-    public ExceptionResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
