--- rdb/src/main/java/com/sequenceiq/rdb/controller/ApiDocController.java
@@ -1,15 +1,0 @@
-package com.sequenceiq.rdb.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-public class ApiDocController {
-
-    @RequestMapping(value = "/apidocs", method = RequestMethod.GET)
-    public ModelAndView apiDocs() {
-        return new ModelAndView("redirect:swagger-ui.html");
-    }
-}
