--- rdb/src/main/java/com/sequenceiq/rdb/controller/mapper/SpringBadRequestExceptionMapper.java
@@ -1,22 +1,0 @@
-package com.sequenceiq.rdb.controller.mapper;
-
-import javax.ws.rs.core.Response.Status;
-
-import org.springframework.stereotype.Component;
-
-import com.sequenceiq.rdb.controller.exception.BadRequestException;
-
-@Component
-public class SpringBadRequestExceptionMapper extends BaseExceptionMapper<BadRequestException> {
-
-    @Override
-    Status getResponseStatus() {
-        return Status.BAD_REQUEST;
-    }
-
-    @Override
-    Class<BadRequestException> getExceptionType() {
-        return BadRequestException.class;
-    }
-
-}
