--- rdb/src/main/java/com/sequenceiq/rdb/controller/mapper/WebApplicationExceptionMapper.java
@@ -1,15 +1,0 @@
-package com.sequenceiq.rdb.controller.mapper;
-
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-public class WebApplicationExceptionMapper implements ExceptionMapper<WebApplicationException> {
-
-    @Override
-    public Response toResponse(WebApplicationException exception) {
-        return exception.getResponse();
-    }
-}
