--- rdb/src/main/java/com/sequenceiq/rdb/service/ha/RdbNodeConfig.java
@@ -1,20 +1,0 @@
-package com.sequenceiq.rdb.service.ha;
-
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-public class RdbNodeConfig {
-
-    @Value("${rdb.instance.node.id:}")
-    private String id;
-
-    public String getId() {
-        return isNodeIdSpecified() ? id : null;
-    }
-
-    public boolean isNodeIdSpecified() {
-        return StringUtils.isNoneBlank(id);
-    }
-}
