--- rdb/src/main/java/com/sequenceiq/rdb/service/registry/ServiceAddressResolver.java
@@ -1,7 +1,0 @@
-package com.sequenceiq.rdb.service.registry;
-
-public interface ServiceAddressResolver {
-    String resolveUrl(String serverUrl, String protocol, String serviceId) throws ServiceAddressResolvingException;
-
-    String resolveHostPort(String host, String port, String serviceId) throws ServiceAddressResolvingException;
-}
