+++ redbeams/src/main/java/com/sequenceiq/redbeams/RedbeamsApplication.java
@@ -1,0 +1,18 @@
+package com.sequenceiq.redbeams;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.actuate.autoconfigure.metrics.web.servlet.WebMvcMetricsAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@EnableSwagger2
+@SpringBootApplication(scanBasePackages = "com.sequenceiq.redbeams", exclude = WebMvcMetricsAutoConfiguration.class)
+public class RedbeamsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RedbeamsApplication.class, args);
+    }
+
+}
+
