+++ redbeams/src/main/java/com/sequenceiq/redbeams/api/RedbeamsApi.java
@@ -1,0 +1,10 @@
+package com.sequenceiq.redbeams.api;
+
+public class RedbeamsApi {
+
+    public static final String API_ROOT_CONTEXT = "/api";
+
+    private RedbeamsApi() {
+    }
+}
+
