+++ redbeams/src/main/java/com/sequenceiq/redbeams/api/model/ExceptionResponse.java
@@ -1,0 +1,18 @@
+package com.sequenceiq.redbeams.api.model;
+
+public class ExceptionResponse {
+
+    private String message;
+
+    public ExceptionResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
