+++ redbeams/src/main/java/com/sequenceiq/redbeams/configuration/CombinedSwaggerResourcesProvider.java
@@ -1,0 +1,36 @@
+package com.sequenceiq.redbeams.configuration;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import javax.annotation.Resource;
+
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Component;
+
+import com.sequenceiq.redbeams.api.RedbeamsApi;
+
+import springfox.documentation.swagger.web.InMemorySwaggerResourcesProvider;
+import springfox.documentation.swagger.web.SwaggerResource;
+import springfox.documentation.swagger.web.SwaggerResourcesProvider;
+
+@Component
+@Primary
+public class CombinedSwaggerResourcesProvider implements SwaggerResourcesProvider {
+
+    @Resource
+    private InMemorySwaggerResourcesProvider inMemorySwaggerResourcesProvider;
+
+    @Override
+    public List<SwaggerResource> get() {
+
+        SwaggerResource jerseySwaggerResource = new SwaggerResource();
+        jerseySwaggerResource.setLocation(RedbeamsApi.API_ROOT_CONTEXT + "/swagger.json");
+        jerseySwaggerResource.setSwaggerVersion("2.0");
+        jerseySwaggerResource.setName("Redbeams API");
+
+        return Stream.concat(Stream.of(jerseySwaggerResource), inMemorySwaggerResourcesProvider.get().stream()).collect(Collectors.toList());
+    }
+
+}
