+++ redbeams/src/main/java/com/sequenceiq/redbeams/controller/exception/BadRequestException.java
@@ -1,0 +1,30 @@
+package com.sequenceiq.redbeams.controller.exception;
+
+import java.util.Objects;
+
+public class BadRequestException extends RuntimeException {
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+
+    public BadRequestException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o instanceof BadRequestException) {
+            return getMessage().equals(((Throwable) o).getMessage());
+        }
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(getMessage());
+    }
+}
