+++ redbeams/src/main/java/com/sequenceiq/redbeams/controller/mapper/DefaultExceptionMapper.java
@@ -1,0 +1,25 @@
+package com.sequenceiq.redbeams.controller.mapper;
+
+import javax.ws.rs.core.Response.Status;
+import javax.ws.rs.ext.Provider;
+
+import com.sequenceiq.redbeams.api.model.ExceptionResponse;
+
+@Provider
+public class DefaultExceptionMapper extends BaseExceptionMapper<Exception> {
+
+    @Override
+    protected Object getEntity(Exception exception) {
+        return new ExceptionResponse("Internal server error: " + exception.getMessage());
+    }
+
+    @Override
+    Status getResponseStatus() {
+        return Status.INTERNAL_SERVER_ERROR;
+    }
+
+    @Override
+    Class<Exception> getExceptionType() {
+        return Exception.class;
+    }
+}
