+++ redbeams/src/main/java/com/sequenceiq/redbeams/controller/mapper/SpringBadRequestExceptionMapper.java
@@ -1,0 +1,22 @@
+package com.sequenceiq.redbeams.controller.mapper;
+
+import javax.ws.rs.core.Response.Status;
+
+import org.springframework.stereotype.Component;
+
+import com.sequenceiq.redbeams.controller.exception.BadRequestException;
+
+@Component
+public class SpringBadRequestExceptionMapper extends BaseExceptionMapper<BadRequestException> {
+
+    @Override
+    Status getResponseStatus() {
+        return Status.BAD_REQUEST;
+    }
+
+    @Override
+    Class<BadRequestException> getExceptionType() {
+        return BadRequestException.class;
+    }
+
+}
