+++ redbeams/src/main/java/com/sequenceiq/redbeams/service/registry/ServiceAddressResolver.java
@@ -1,0 +1,7 @@
+package com.sequenceiq.redbeams.service.registry;
+
+public interface ServiceAddressResolver {
+    String resolveUrl(String serverUrl, String protocol, String serviceId) throws ServiceAddressResolvingException;
+
+    String resolveHostPort(String host, String port, String serviceId) throws ServiceAddressResolvingException;
+}
