+++ redbeams/src/main/java/com/sequenceiq/redbeams/service/registry/ServiceAddressResolvingException.java
@@ -1,0 +1,15 @@
+package com.sequenceiq.redbeams.service.registry;
+
+public class ServiceAddressResolvingException extends Exception {
+    public ServiceAddressResolvingException(String message) {
+        super(message);
+    }
+
+    public ServiceAddressResolvingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ServiceAddressResolvingException(Throwable cause) {
+        super(cause);
+    }
+}
