+++ settings.gradle
@@ -18,7 +18,7 @@ include 'cloud-yarn'
 include 'cloud-cumulus-yarn'
 include 'core-api'
 include 'datalake'
-include 'rdb'
+include 'redbeams'
 include 'autoscale'
 include 'autoscale-api'
 include 'structuredevent-model'
